The code in this directory should be used as reference code.
TDah is the only folder that is known to work from top to bottom and as of today will be the only folder in the trunk.  All other files/projects may not work.  If they do not work, then it should not take too long to make them work again, errors may include, but not be limited to, setting up the correct libraries commenting/uncommenting certain lines, etc.  Many of the functions in these files do work by themselves even if the entire project does not compile.